+++ src/peripheral/ch32v00x/field/rcc/ctlr.h
@@ -23,7 +23,7 @@ namespace peripheral::rcc {
             field_bit_mask = 0b1U,         // not holding any settings or value, it's a bitmask for this specific field
             field_access   = soc::reg::field_access_right::ReadWrite,
 
-            disable        = 0U,           // after disabling HSI, it takes 6 cycles to propagate the change
+            disable        = 0U,             // after disabling HSI, it takes 6 cycles to propagate the change
             enable         = field_bit_mask, // HSI can set by HW as well (when waking up from standby or when experiencing failure while using HSE). v003=24Mhz, x033/x035=48Mhz others=8Mhz
         };
 
@@ -31,44 +31,44 @@ namespace peripheral::rcc {
             field_bit_mask   = 0b1U << 1U,    // not holding any settings or value, it's a bitmask for this specific field
             field_access     = soc::reg::field_access_right::ReadOnly,
 
-            not_ready        = 0U,           // HSI not ready/stable yet, wait for a bit longer
+            not_ready        = 0U,             // HSI not ready/stable yet, wait for a bit longer
             ready_and_stable = field_bit_mask, // HSI can set by HW as well (when waking up from standby or when experiencing failure while using HSE)
         };
 
         enum class HSITRIM_RW_InternalHighSpeedClockTrim: std::uint32_t { // NOLINT
             // Able to superimpose on top of HSICAL[7:0] to adjust HSI frequency. Range is 0-31 60kHz steps, value 16 is the default midpoint
-            // HSITRIM=0  coresponds to -960 kHz
-            // HSITRIM=1  coresponds to -900 kHz
-            // HSITRIM=2  coresponds to -840 kHz
-            // HSITRIM=3  coresponds to -780 kHz
-            // HSITRIM=4  coresponds to -720 kHz
-            // HSITRIM=5  coresponds to -660 kHz
-            // HSITRIM=6  coresponds to -600 kHz
-            // HSITRIM=7  coresponds to -540 kHz
-            // HSITRIM=8  coresponds to -480 kHz
-            // HSITRIM=9  coresponds to -420 kHz
-            // HSITRIM=10 coresponds to -360 kHz
-            // HSITRIM=11 coresponds to -300 kHz
-            // HSITRIM=12 coresponds to -240 kHz
-            // HSITRIM=13 coresponds to -180 kHz
-            // HSITRIM=14 coresponds to -120 kHz
-            // HSITRIM=15 coresponds to  -60 kHz
-            // HSITRIM=16 coresponds to    0 kHz
-            // HSITRIM=17 coresponds to   60 kHz
-            // HSITRIM=18 coresponds to  120 kHz
-            // HSITRIM=19 coresponds to  180 kHz
-            // HSITRIM=20 coresponds to  240 kHz
-            // HSITRIM=21 coresponds to  300 kHz
-            // HSITRIM=22 coresponds to  360 kHz
-            // HSITRIM=23 coresponds to  420 kHz
-            // HSITRIM=24 coresponds to  480 kHz
-            // HSITRIM=25 coresponds to  540 kHz
-            // HSITRIM=26 coresponds to  600 kHz
-            // HSITRIM=27 coresponds to  660 kHz
-            // HSITRIM=28 coresponds to  720 kHz
-            // HSITRIM=29 coresponds to  780 kHz
-            // HSITRIM=30 coresponds to  840 kHz
-            // HSITRIM=31 coresponds to  900 kHz
+            // HSITRIM=0  corresponds to -960 kHz
+            // HSITRIM=1  corresponds to -900 kHz
+            // HSITRIM=2  corresponds to -840 kHz
+            // HSITRIM=3  corresponds to -780 kHz
+            // HSITRIM=4  corresponds to -720 kHz
+            // HSITRIM=5  corresponds to -660 kHz
+            // HSITRIM=6  corresponds to -600 kHz
+            // HSITRIM=7  corresponds to -540 kHz
+            // HSITRIM=8  corresponds to -480 kHz
+            // HSITRIM=9  corresponds to -420 kHz
+            // HSITRIM=10 corresponds to -360 kHz
+            // HSITRIM=11 corresponds to -300 kHz
+            // HSITRIM=12 corresponds to -240 kHz
+            // HSITRIM=13 corresponds to -180 kHz
+            // HSITRIM=14 corresponds to -120 kHz
+            // HSITRIM=15 corresponds to  -60 kHz
+            // HSITRIM=16 corresponds to    0 kHz
+            // HSITRIM=17 corresponds to   60 kHz
+            // HSITRIM=18 corresponds to  120 kHz
+            // HSITRIM=19 corresponds to  180 kHz
+            // HSITRIM=20 corresponds to  240 kHz
+            // HSITRIM=21 corresponds to  300 kHz
+            // HSITRIM=22 corresponds to  360 kHz
+            // HSITRIM=23 corresponds to  420 kHz
+            // HSITRIM=24 corresponds to  480 kHz
+            // HSITRIM=25 corresponds to  540 kHz
+            // HSITRIM=26 corresponds to  600 kHz
+            // HSITRIM=27 corresponds to  660 kHz
+            // HSITRIM=28 corresponds to  720 kHz
+            // HSITRIM=29 corresponds to  780 kHz
+            // HSITRIM=30 corresponds to  840 kHz
+            // HSITRIM=31 corresponds to  900 kHz
 
             field_bit_mask = 0b11111U << 3U, // not holding any settings or value, it's a bitmask for this specific field
             field_access   = soc::reg::field_access_right::ReadWrite,
@@ -77,59 +77,59 @@ namespace peripheral::rcc {
         template<std::uint8_t TplTrim>
         constexpr static auto produceRawTrimValueCt() -> std::uint32_t {
             static_assert(TplTrim < 32U, "TplTrim is 5-bits, the value must be in range 0-31");
-            return TplTrim << soc::reg::enum_bit_mask_offset_ct<HSITRIM_RW_InternalHighSpeedClockTrim>();
+            return static_cast<std::uint32_t>(TplTrim) << static_cast<std::uint32_t>(soc::reg::enum_bit_mask_offset_ct<HSITRIM_RW_InternalHighSpeedClockTrim>());
         }
 
         enum class HSICAL_RO_InternalHighSpeedClockCalibration: std::uint32_t { // NOLINT
-            field_bit_mask = 0b11111111U << 8U, // not holding any settings or value, it's a bitmask for this specific field
+            field_bit_mask = static_cast<std::uint32_t>(0b11111111U) << 8U, // not holding any settings or value, it's a bitmask for this specific field
             field_access   = soc::reg::field_access_right::ReadOnly,
         };
 
         enum class HSEON_RW_ExternalHighSpeedClockEnable: std::uint32_t { // NOLINT
-            field_bit_mask = 0b1U << 16U,   // not holding any settings or value, it's a bitmask for this specific field
+            field_bit_mask = static_cast<std::uint32_t>(0b1U) << 16U,   // not holding any settings or value, it's a bitmask for this specific field
             field_access   = soc::reg::field_access_right::ReadWrite,
 
-            disable        = 0U,           // default, changes take 6 cycles to apply
+            disable        = 0U,             // default, changes take 6 cycles to apply
             enable         = field_bit_mask, // changes take 6 cycles to apply. turned off when entering standby mode,
         };
 
         enum class HSERDY_RO_ExternalHighSpeedClockReady: std::uint32_t { // NOLINT
-            field_bit_mask   = 0b1U << 17U,  // not holding any settings or value, it's a bitmask for this specific field
+            field_bit_mask   = static_cast<std::uint32_t>(0b1U) << 17U,  // not holding any settings or value, it's a bitmask for this specific field
             field_access     = soc::reg::field_access_right::ReadOnly,
 
-            not_ready        = 0U,           // default, HSE not ready/stable yet, wait for a bit longer
+            not_ready        = 0U,             // default, HSE not ready/stable yet, wait for a bit longer
             ready_and_stable = field_bit_mask, // HSE will be turned off when entering standby mode
         };
 
         enum class HSEBYP_RW_ExternalHighSpeedClockBypass: std::uint32_t { // NOLINT
-            field_bit_mask           = 0b1U << 18U,  // not holding any settings or value, it's a bitmask for this specific field
+            field_bit_mask           = static_cast<std::uint32_t>(0b1U) << 18U,  // not holding any settings or value, it's a bitmask for this specific field
             field_access             = soc::reg::field_access_right::ReadWrite,
 
-            use_ceramic_resonator    = 0U,           // default, resonate oscillator connected to the external pins, configure this while HSEON is off
+            use_ceramic_resonator    = 0U,             // default, resonate oscillator connected to the external pins, configure this while HSEON is off
             bypass_ceramic_resonator = field_bit_mask, // consume digital clock signal on external pin, configure this while HSEON is off
         };
 
         enum class CSSON_RW_ClockSafety: std::uint32_t { // NOLINT
-            field_bit_mask = 0b1U << 19U,  // not holding any settings or value, it's a bitmask for this specific field
+            field_bit_mask = static_cast<std::uint32_t>(0b1U) << 19U,  // not holding any settings or value, it's a bitmask for this specific field
             field_access   = soc::reg::field_access_right::ReadWrite,
 
-            no_protection  = 0U,           // disable protection HSE issues will not cause IRQ or flags to be set
+            no_protection  = 0U,             // disable protection HSE issues will not cause IRQ or flags to be set
             protect_hse    = field_bit_mask, // when HSERDY=1 triggers CSSF and NMI IRQ when HSE abnormal, when HSERDY=0 disable protection
         };
 
         enum class PLLON_RW_PhaseLockedLoopEnable: std::uint32_t { // NOLINT
-            field_bit_mask = 0b1U << 24U,  // not holding any settings or value, it's a bitmask for this specific field
+            field_bit_mask = static_cast<std::uint32_t>(0b1U) << 24U,  // not holding any settings or value, it's a bitmask for this specific field
             field_access   = soc::reg::field_access_right::ReadWrite,
 
-            disable        = 0U,           // clock generator is off
+            disable        = 0U,             // clock generator is off
             enable         = field_bit_mask, // clock generator is on, but will be disabled when entering standby
         };
 
         enum class PLLRDY_RO_PhaseLockedLoopReady: std::uint32_t { // NOLINT
-            field_bit_mask   = 0b1U << 25U,  // not holding any settings or value, it's a bitmask for this specific field
+            field_bit_mask   = static_cast<std::uint32_t>(0b1U) << 25U,  // not holding any settings or value, it's a bitmask for this specific field
             field_access     = soc::reg::field_access_right::ReadOnly,
 
-            not_ready        = 0U,           // PLL clock lock NOT achieved yet, wait for a bit longer
+            not_ready        = 0U,             // PLL clock lock NOT achieved yet, wait for a bit longer
             ready_and_stable = field_bit_mask, // PLL clock lock achieved
         };
